Fix QNanoPainter OpenGL context usage
When using multiple QNanoWidgets, each widget will get a unique instance
of a QOpenGLContext object. When the global instance of QNanoPainter is
initialised, it happens against the OpenGL context owned by one of the
QNanoWidgets (which ever is painted first).

This means that the global instance of QNanoPainter may not work correctly
when used with other QNanoWidget instances. There are 2 things needed to
fix this:

- The app must enable the Qt::AA_ShareOpenGLContexts attribute.
  This allows the QNanoPainter to be initialised against one
  QOpenGLContext, but also used with other QOpenGLContexts.

- The QNanoPainter should maintain its own QOpenGLContext instance, and
  ensure the QNanoPainter backend is initialised against that. This is
  because, on Windows, the QNanoPainter backend retrieves the OpenGL
  function pointers from the active QOpenGLContext, but then uses them as
  a global variable. This means that once that particular QOpenGLContext
  is destroyed (ie. the QNanoWidget is destroyed), the OpenGL function
  pointers are also destroyed, but the global QNanoPainter backend still
  continues to reference them.